+++ core/class/prixcarburants.class.php
@@ -47,20 +47,24 @@ class prixcarburants extends eqLogic {
 		return round(($earth_radius * $d)/1000);
 	}
 
-
 	public static function custom_sort($a,$b) {
 	   return $a['prix']>$b['prix'];
 	}
 
-	public static function getMarqueStation($idstation) {
-		$json = file_get_contents(__DIR__."/stations.json");
-		$parsed_json = json_decode($json, true);
-		foreach($parsed_json as $row) {
-			if($row['id'] == $idstation) {
-				return $row['marque'];
-				break;
-			}
-		}
+	public static function getMarqueStation($idstation, $DepStation) {
+	    $json = @file_get_contents(__DIR__ . '/listestations/stations' . $DepStation . '.json');
+	    if($json!==false){
+    		$parsed_json = json_decode($json, true);
+    		foreach($parsed_json['stations'] as $row) {
+    			if($row['id'] == $idstation) {
+    				return $row['marque'];
+    				break;
+    			}
+    		}
+	    } else {
+	        log::add('prixcarburants','debug','JSON file : /listestations/stations' . $DepStation . '.json not available');
+	        return __('Erreur',  __FILE__);
+	    }
 	}
 
 	public static function MAJVehicules($oneveh) {
@@ -84,88 +88,127 @@ class prixcarburants extends eqLogic {
 			$nom = $unvehicule->getName();
 			$typecarburant = $unvehicule->getConfiguration('typecarburant','');
 			$rayon = $unvehicule->getConfiguration('rayon','30');
-			$station1 = $unvehicule->getConfiguration('station1','');
-			$monformatdate = $unvehicule->getConfiguration('formatdate','');
 			$nbstation = $unvehicule->getConfiguration('nbstation','3');
-			
-			//Get position latitude and longitude only if favorite station is not selected
-			if ($station1 == ''){
+			$monformatdate = $unvehicule->getConfiguration('formatdate','');
+			//Get the list of favoris selected
+			$NbFavoris = 0;
+			if ($unvehicule->getConfiguration('Favoris') == '1') {
+    			for($i = 1; $i <=10; $i++) {
+    			    if($unvehicule->getConfiguration('station' . $i . '_Dep','') != '' && $unvehicule->getConfiguration('station' . $i . '_Station','') != '') {
+    			        $StationFav[$i] = $unvehicule->getConfiguration('station' . $i . '_Station','');
+    			        $DepartementFav[$i] = $unvehicule->getConfiguration('station' . $i . '_Dep','');
+    			        $NbFavoris++;
+    			    } else {
+    			        break;
+    			    }
+    			}
+			}
+			//Get position latitude and longitude only if geolocalisation is selected
+			if ($unvehicule->getConfiguration('ViaLoca') == '1') {
     			if ($unvehicule->getConfiguration('geoloc', 'none') == 'none') {
     			    $macmd = cmd::byEqLogicIdCmdName($unvehicule->getId(),'Top 1 Adresse');
     			    if (is_object($macmd)) $macmd->event(__('Pas de localisation sélectionnée',  __FILE__));
-    			    //if (is_object($macmd)) $macmd->event('{{Pas de localisation sélectionnée}}');
     			    return;
     			} elseif ($unvehicule->getConfiguration('geoloc') == "jeedom") {
     			    $malat = config::byKey('info::latitude');
     			    $malng = config::byKey('info::longitude');
     			} else {
-    			    $coordonnees = geotravCmd::byEqLogicIdAndLogicalId($unvehicule->getConfiguration('geoloc'),'location:coordinate')->execCmd();
+    			    if (geotravCmd::byEqLogicIdAndLogicalId($unvehicule->getConfiguration('geoloc'),'location:coordinate') != null){
+                        $coordonnees = geotravCmd::byEqLogicIdAndLogicalId($unvehicule->getConfiguration('geoloc'),'location:coordinate')->execCmd();
+    			    }else{
+    			        $coordonnees = cmd::byId($unvehicule->getConfiguration('geoloc'))->execCmd();
+    			    }
     			    $expcoord = explode(",",$coordonnees);
     		        $malat = $expcoord[0];
     		        $malng = $expcoord[1];
     			}
 			}
 			
+			//Parse XML file
 			while($reader->read()) {
 				if ($reader->nodeType == XMLReader::ELEMENT && $reader->name == 'pdv') {
 					$lat = $reader->getAttribute('latitude')/100000;
 					$lng = $reader->getAttribute('longitude')/100000;
 					$mastationid = $reader->getAttribute('id');
+					$MaStationDep = intval(substr($reader->getAttribute('cp'), 0, 2));
 					$dist = 0;
+					$MonTest = False;
+					$EstFavoris = False;
 					
-					if ($station1 == ''){
-						$dist = prixcarburants::distance($malat,$malng,$lat,$lng);
-						if ($dist > $rayon) continue; //On enregistrer uniquement ceux qui sont a proximitées
+					//Check if this station is a favorite
+					if($NbFavoris > 0) {
+					    for($i = 1; $i <= $NbFavoris; $i++) {
+					        if($mastationid == $StationFav[$i] && $MaStationDep == $DepartementFav[$i]) {
+					            $MonTest = True;
+					            $EstFavoris = True;
+					            $ordreFav = $i - 1;
+					            break;
+					        }
+					    }
 					}
+					if($MonTest == False) {
+    					if($unvehicule->getConfiguration('ViaLoca') == '1'){
+    						$dist = prixcarburants::distance($malat,$malng,$lat,$lng);
+    						if($dist <= $rayon) $MonTest = True;
+    					}
+					}
+					
+					//Register only station that are a favorite or on max radius
+					if($MonTest == False) continue;
 					
 					$unestation = simplexml_import_dom($doc->importNode($reader->expand(), true));
 					
 					foreach ($unestation->prix as $prix){
-						if ($prix->attributes()->nom == $typecarburant){ // FILTRER SELON CARBURANTS
-						
-							if ($station1 != '') {
-								if ($mastationid != $station1) continue; 
-							}
-							
+						if ($prix->attributes()->nom == $typecarburant){ //Filter by fuel type
 							$prixlitre = $prix->attributes()->valeur.'';
 							$maj = $prix->attributes()->maj.'';
-							$marque = prixcarburants::getMarqueStation($mastationid);
+							$marque = prixcarburants::getMarqueStation($mastationid, $MaStationDep);
 							
-							$maselection[$idx]['adresse'] = $marque.', '.$unestation->ville;
-							$maselection[$idx]['prix'] = $prixlitre;
-							$maselection[$idx]['maj']  = date($monformatdate, strtotime($maj));
-							$maselection[$idx]['distance'] = $dist;
-							$maselection[$idx]['id'] = $mastationid;
-							$idx++;
+							if($EstFavoris) {
+							    $SelectionFav[$ordreFav]['adresse'] = $marque.', '.$unestation->ville;
+							    $SelectionFav[$ordreFav]['prix'] = $prixlitre;
+							    $SelectionFav[$ordreFav]['maj']  = date($monformatdate, strtotime($maj));
+							    $SelectionFav[$ordreFav]['distance'] = $dist;
+							    $SelectionFav[$ordreFav]['id'] = $mastationid;
+							} else {
+							    $maselection[$idx]['adresse'] = $marque.', '.$unestation->ville;
+							    $maselection[$idx]['prix'] = $prixlitre;
+							    $maselection[$idx]['maj']  = date($monformatdate, strtotime($maj));
+							    $maselection[$idx]['distance'] = $dist;
+							    $maselection[$idx]['id'] = $mastationid;
+							    $idx++;
+							}
 						}
 					}
-				}//PDV
-			}// XML
+				}
+			}
 			$reader->close();
 
 			log::add('prixcarburants','debug',' step count selection '.count($maselection).' '.$nom);
-
-			if ($station1 != '') $nbstation = 1; //If a favorite is selected, only 1 station need to be fill
 				
 			usort($maselection, "prixcarburants::custom_sort");
 			
 			For($i = 1; $i <= $nbstation; $i++) {
-			    if($maselection[$i - 1]['prix'] != ''){ //Price available so, there is enough station on selected range
+			    if($i <= $NbFavoris) {
+			        $liste[$i - 1] = $SelectionFav[$i - 1];
+			    } else {
+			        $liste[$i - 1] = $maselection[$i - 1 - $NbFavoris];
+			    }
+			    if($liste[$i - 1]['prix'] != ''){ //Price available so, there is enough station on selected range
     				$macmd = cmd::byEqLogicIdCmdName($unvehicule->getId(),'Top ' . $i . ' ID');
-    				if (is_object($macmd)) $macmd->event($maselection[$i - 1]['id']);
+    				if (is_object($macmd)) $macmd->event($liste[$i - 1]['id']);
     				
     				$macmd = cmd::byEqLogicIdCmdName($unvehicule->getId(),'Top ' . $i . ' Adresse');
-    				if (is_object($macmd)) $macmd->event($maselection[$i - 1]['adresse']);
+    				if (is_object($macmd)) $macmd->event($liste[$i - 1]['adresse']);
     				
     				$macmd = cmd::byEqLogicIdCmdName($unvehicule->getId(),'Top ' . $i . ' MAJ');
-    				if (is_object($macmd)) $macmd->event($maselection[$i - 1]['maj']);
+    				if (is_object($macmd)) $macmd->event($liste[$i - 1]['maj']);
     				
     				$macmd = cmd::byEqLogicIdCmdName($unvehicule->getId(),'Top ' . $i . ' Prix');
-    				if (is_object($macmd)) $macmd->event($maselection[$i - 1]['prix']);
+    				if (is_object($macmd)) $macmd->event($liste[$i - 1]['prix']);
 			    } else {
 			        $macmd = cmd::byEqLogicIdCmdName($unvehicule->getId(),'Top ' . $i . ' Adresse');
 			        if (is_object($macmd)) $macmd->event(__('Plus de station disponible dans le rayon sélectionné',  __FILE__));
-			        //if (is_object($macmd)) $macmd->event('{{Plus de station disponible dans le rayon sélectionné}}');
 			    }
 			}
 			
@@ -182,16 +225,15 @@ class prixcarburants extends eqLogic {
 		
 		$zip = new ZipArchive;
 		if ($zip->open($filenamezip) === TRUE) {
-		$zip->extractTo(__DIR__);
-		$zip->close();
-		log::add('prixcarburants','debug','prix zip ok get'.__DIR__);
-		
-		unlink(__DIR__.'/PrixCarburants.zip');
-		
-		prixcarburants::MAJVehicules(null);
-		
+    		$zip->extractTo(__DIR__);
+    		$zip->close();
+    		log::add('prixcarburants','debug','prix zip ok get'.__DIR__);
+    		
+    		unlink(__DIR__.'/PrixCarburants.zip');
+    		
+    		prixcarburants::MAJVehicules(null);
 		} else {
-		log::add('prixcarburants','debug','prix zip nok get'.__DIR__);
+		  log::add('prixcarburants','debug','prix zip nok get'.__DIR__);
 		}
 	}
 	
@@ -221,7 +263,10 @@ class prixcarburants extends eqLogic {
 
 	public function preInsert() {}
 
-	public function postInsert() {}
+	public function postInsert() {
+	    //Update prixcarburants file
+	    //$this->getEqLogic()->updatePrixCarburant();
+	}
 
 	public function preSave() {}
 
@@ -231,7 +276,7 @@ class prixcarburants extends eqLogic {
 		if (!is_object($prixcarburantsCmd)) {
 			log::add('prixcarburants', 'debug', 'refresh');
 			$prixcarburantsCmd = new prixcarburantsCmd();
-			$prixcarburantsCmd->setName(__('Rafraichir',  __FILE__));
+			$prixcarburantsCmd->setName(__('Actualiser',  __FILE__));
 			$prixcarburantsCmd->setEqLogic_id($this->getId());
 			$prixcarburantsCmd->setLogicalId('refresh');
 			$prixcarburantsCmd->setType('action');
@@ -240,6 +285,11 @@ class prixcarburants extends eqLogic {
 			$prixcarburantsCmd->setIsVisible(0);
 			$prixcarburantsCmd->setOrder(0);
 			$prixcarburantsCmd->save();
+		}
+		
+		//Create file with fuel price if it doesn't exist (first creation)
+		if (!file_exists(__DIR__.'/PrixCarburants_instantane.xml')) {
+		  $prixcarburantsCmd->getEqLogic()->updatePrixCarburant();
 		}
 	}