+++ docs/en_US/changelog.md
@@ -1,11 +1,24 @@
 # Changelog : Prix Carburants
 
 ### TODO list
-- Create a full documenation
-- Update the file containing gaz station list (stations.json)
-- Improve the selection of favorite gaz station
-- Add the possibility on a vehicle to follow a favorite and see station arround a location, at the same time
 - Add coloration to highlight data that are up to date in comparison with data that has not been updated since a while
+
+## 03/07/2020
+- Mise à jour des fichiers contenant la liste des stations (JSON)
+- Améliorer la sélection de la station favoris
+- Ajout de la possibilité, sur un seul véhicule avoir le choix entre suivre une station favorite et des stations autour des coordonnées
+- When the first equipement is created, creation of the list of price.
+
+## 01/07/2020
+- Show only activated geoloc
+However if you have the message "No more stations available in the selected radius", please select again the location option and save the equipement.
+
+## 30 Juin 2020
+- Add dynamix geolocation through Geoloc plugin (for car in mouvement, for example)
+
+## 24 juin 2020
+- Documentation update
+- Added on Jeedom market !
 
 ## 16 juin 2020
 - Translation created (fr_FR and en_US)